Set up clustering notebook
